Install command and cache refactor
